Completed App
In this front end app contains completed full functionality

+++ src/app/dashboard/category/category.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { AddCategoryComponent } from '../add-category/add-category.component';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ApicallsService } from '../services/apicalls.service';
+import { DeleteAlertComponent } from '../delete-alert/delete-alert.component';
+import { CantDeleteAlertComponent } from '../cant-delete-alert/cant-delete-alert.component';
 @Component({
   selector: 'app-category',
   templateUrl: './category.component.html',
@@ -23,7 +25,7 @@ export class CategoryComponent {
   }
   fetchDetails() {
     this.userId = localStorage.getItem('userId');
-    this.apiService.getAllCategory(this.userId)
+    this.apiService.getAllCategory()
       .subscribe(x => {
         console.log("Get Api data", x)
         this.tableData = x.response;
@@ -43,9 +45,19 @@ export class CategoryComponent {
     })
   }
   deleteRow(id: String) {
-    this.apiService.deleteCategory(id)
-    .subscribe(x => {
-      this.fetchDetails();
+    const model = this.model.open(DeleteAlertComponent, { size: 'md' })
+    model.result.then(r => {
+      if (r) {
+        this.apiService.deleteCategory(id)
+          .subscribe(x => {
+            console.log("Printing delete response",x)
+            if(x.res == 0){
+              const popup = this.model.open(CantDeleteAlertComponent,{size : 'md'})
+              popup.componentInstance.content = "Category contains Sub-Category data not able to delete!!!";
+            }
+            this.fetchDetails();
+          })
+      }
     })
   }
   addCategory() {
